Removed module for getOppositeDirection and moved it to the findChargePath module, the only module requiring it. #28
