--- client/utils/get-opposite-direction.js
@@ -1,20 +1,0 @@
-function getOppositeDirection(validChannel) {
-  let channelOpposite = null
-  switch (validChannel) {
-    case 0:
-      channelOpposite = 'south'
-      break
-    case 1:
-      channelOpposite = 'north'
-      break
-    case 2:
-      channelOpposite = 'west'
-      break
-    case 3:
-      channelOpposite = 'east'
-      break
-  }
-  return channelOpposite
-}
-
-module.exports = getOppositeDirection
